digitalocean-scheduled/packages/narrativa-manager/scheduled/requirements.txt
requests==2.32.3
redis==6.2.0
python-dotenv